requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
colorama==0.4.6
    # via -r requirements.in
pillow==10.3.0
    # via -r requirements.in
